requirements.txt
Appium-Python-Client==3.1.1
pytest==8.1.1
selenium==4.20.0
